consistent message persisting by removing dependence on index

+++ core/message/msgservice/model.go
@@ -3,7 +3,7 @@ package msgservice
 import "fmt"
 
 type Message struct {
-	Offset int
+	Offset uint64
 	Body   []byte
 }
